+++ src/podofo/main/PdfEncrypt.cpp
@@ -12,6 +12,8 @@
 #include <podofo/private/PdfDeclarationsPrivate.h>
 #include "PdfEncrypt.h"
 
+#include <podofo/private/OpenSSLInternal.h>
+
 #include "PdfDictionary.h"
 #include "PdfFilter.h"
 
@@ -44,74 +46,6 @@ PdfEncryptAlgorithm::RC4V1 |
 PdfEncryptAlgorithm::RC4V2 |
 PdfEncryptAlgorithm::AESV2;
 #endif // PODOFO_HAVE_LIBIDN
-
-#if OPENSSL_VERSION_MAJOR >= 3
-#include <openssl/provider.h>
-#endif // OPENSSL_VERSION_MAJOR >= 3
-
-class OpenSSLInit
-{
-public:
-    OpenSSLInit()
-    {
-#if OPENSSL_VERSION_MAJOR >= 3
-        m_libCtx = OSSL_LIB_CTX_new();
-        if (m_libCtx == nullptr)
-            PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InvalidHandle, "Unable to create OpenSSL library context");
-
-        // NOTE: Load required legacy providers, such as RC4, together regular ones,
-        // as explained in https://wiki.openssl.org/index.php/OpenSSL_3.0#Providers
-        m_legacyProvider = OSSL_PROVIDER_load(m_libCtx, "legacy");
-        if (m_legacyProvider == nullptr)
-            PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InvalidHandle, "Unable to load legacy providers in OpenSSL >= 3.x.x");
-
-        m_defaultProvider = OSSL_PROVIDER_load(m_libCtx, "default");
-        if (m_defaultProvider == nullptr)
-            PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InvalidHandle, "Unable to load default providers in OpenSSL >= 3.x.x");
-
-        // https://www.openssl.org/docs/man3.0/man7/crypto.html#FETCHING-EXAMPLES
-        Rc4 = EVP_CIPHER_fetch(m_libCtx, "RC4", "provider=legacy");
-        Aes128 = EVP_CIPHER_fetch(m_libCtx, "AES-128-CBC", "provider=default");
-        Aes256 = EVP_CIPHER_fetch(m_libCtx, "AES-256-CBC", "provider=default");
-        MD5 = EVP_MD_fetch(m_libCtx, "MD5", "provider=default");
-        SHA256 = EVP_MD_fetch(m_libCtx, "SHA2-256", "provider=default");
-        SHA384 = EVP_MD_fetch(m_libCtx, "SHA2-384", "provider=default");
-        SHA512 = EVP_MD_fetch(m_libCtx, "SHA2-512", "provider=default");
-
-#else // OPENSSL_VERSION_MAJOR < 3
-        Rc4 = EVP_rc4();
-        Aes128 = EVP_aes_128_cbc();
-        Aes256 = EVP_aes_256_cbc();
-        MD5 = EVP_md5();
-        SHA256 = EVP_sha256();
-        SHA384 = EVP_sha384();
-        SHA512 = EVP_sha512();
-#endif // OPENSSL_VERSION_MAJOR >= 3
-    }
-
-    ~OpenSSLInit()
-    {
-#if OPENSSL_VERSION_MAJOR >= 3
-        OSSL_PROVIDER_unload(m_legacyProvider);
-        OSSL_PROVIDER_unload(m_defaultProvider);
-        OSSL_LIB_CTX_free(m_libCtx);
-#endif // OPENSSL_VERSION_MAJOR >= 3
-    }
-public:
-    const EVP_CIPHER* Rc4;
-    const EVP_CIPHER* Aes128;
-    const EVP_CIPHER* Aes256;
-    const EVP_MD* MD5;
-    const EVP_MD* SHA256;
-    const EVP_MD* SHA384;
-    const EVP_MD* SHA512;
-private:
-#if OPENSSL_VERSION_MAJOR >= 3
-    OSSL_LIB_CTX *m_libCtx;
-    OSSL_PROVIDER *m_legacyProvider;
-    OSSL_PROVIDER *m_defaultProvider;
-#endif // OPENSSL_VERSION_MAJOR >= 3
-} s_SSL;
 
 // Default value for P (permissions) = no permission
 #define PERMS_DEFAULT (PdfPermissions)0xFFFFF0C0
@@ -356,13 +290,13 @@ protected:
             {
                 case (size_t)PdfKeyLength::L128 / 8:
                 {
-                    cipher = s_SSL.Aes128;
+                    cipher = ssl::Aes128();
                     break;
                 }
 #ifdef PODOFO_HAVE_LIBIDN
                 case (size_t)PdfKeyLength::L256 / 8:
                 {
-                    cipher = s_SSL.Aes256;
+                    cipher = ssl::Aes256();
                     break;
                 }
 #endif
@@ -711,7 +645,7 @@ void PdfEncryptMD5Base::ComputeOwnerKey(const unsigned char userPad[32], const u
     int rc;
 
     unique_ptr<EVP_MD_CTX, decltype(&EVP_MD_CTX_free)> ctx(EVP_MD_CTX_new(), EVP_MD_CTX_free);
-    if (ctx == nullptr || (rc = EVP_DigestInit_ex(ctx.get(), s_SSL.MD5, nullptr)) != 1)
+    if (ctx == nullptr || (rc = EVP_DigestInit_ex(ctx.get(), ssl::MD5(), nullptr)) != 1)
         PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InternalLogic, "Error initializing MD5 hashing engine");
 
     rc = EVP_DigestUpdate(ctx.get(), ownerPad, 32);
@@ -727,7 +661,7 @@ void PdfEncryptMD5Base::ComputeOwnerKey(const unsigned char userPad[32], const u
         // only use for the input as many bit as the key consists of
         for (int k = 0; k < 50; ++k)
         {
-            rc = EVP_DigestInit_ex(ctx.get(), s_SSL.MD5, nullptr);
+            rc = EVP_DigestInit_ex(ctx.get(), ssl::MD5(), nullptr);
             if (rc != 1)
                 PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InternalLogic, "Error initializing MD5 hashing engine");
 
@@ -769,7 +703,7 @@ void PdfEncryptMD5Base::ComputeEncryptionKey(const string_view& documentId,
     int rc;
 
     unique_ptr<EVP_MD_CTX, decltype(&EVP_MD_CTX_free)> ctx(EVP_MD_CTX_new(), EVP_MD_CTX_free);
-    if (ctx == nullptr || (rc = EVP_DigestInit_ex(ctx.get(), s_SSL.MD5, nullptr)) != 1)
+    if (ctx == nullptr || (rc = EVP_DigestInit_ex(ctx.get(), ssl::MD5(), nullptr)) != 1)
         PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InternalLogic, "Error initializing MD5 hashing engine");
 
     rc = EVP_DigestUpdate(ctx.get(), userPad, 32);
@@ -823,7 +757,7 @@ void PdfEncryptMD5Base::ComputeEncryptionKey(const string_view& documentId,
     {
         for (k = 0; k < 50; ++k)
         {
-            rc = EVP_DigestInit_ex(ctx.get(), s_SSL.MD5, nullptr);
+            rc = EVP_DigestInit_ex(ctx.get(), ssl::MD5(), nullptr);
             if (rc != 1)
                 PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InternalLogic, "Error initializing MD5 hashing engine");
 
@@ -842,7 +776,7 @@ void PdfEncryptMD5Base::ComputeEncryptionKey(const string_view& documentId,
     // Setup user key
     if (revision == 3 || revision == 4)
     {
-        rc = EVP_DigestInit_ex(ctx.get(), s_SSL.MD5, nullptr);
+        rc = EVP_DigestInit_ex(ctx.get(), ssl::MD5(), nullptr);
         if (rc != 1)
             PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InternalLogic, "Error initializing MD5 hashing engine");
 
@@ -935,7 +869,7 @@ void PdfEncryptRC4Base::RC4(const unsigned char* key, unsigned keylen,
         PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InternalLogic, "Error initializing RC4 encryption engine");
 
     // Don't set the key because we will modify the parameters
-    int status = EVP_EncryptInit_ex(rc4, s_SSL.Rc4, nullptr, nullptr, nullptr);
+    int status = EVP_EncryptInit_ex(rc4, ssl::Rc4(), nullptr, nullptr, nullptr);
     if (status != 1)
         PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InternalLogic, "Error initializing RC4 encryption engine");
 
@@ -962,7 +896,7 @@ void PdfEncryptMD5Base::GetMD5Binary(const unsigned char* data, unsigned length,
 {
     int rc;
     unique_ptr<EVP_MD_CTX, decltype(&EVP_MD_CTX_free)> ctx(EVP_MD_CTX_new(), EVP_MD_CTX_free);
-    if (ctx == nullptr || (rc = EVP_DigestInit_ex(ctx.get(), s_SSL.MD5, nullptr)) != 1)
+    if (ctx == nullptr || (rc = EVP_DigestInit_ex(ctx.get(), ssl::MD5(), nullptr)) != 1)
         PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InternalLogic, "Error initializing MD5 hashing engine");
 
     rc = EVP_DigestUpdate(ctx.get(), data, length);
@@ -1229,10 +1163,10 @@ void PdfEncryptAESBase::BaseDecrypt(const unsigned char* key, unsigned keyLen, c
 
     int rc;
     if (keyLen == (int)PdfKeyLength::L128 / 8)
-        rc = EVP_DecryptInit_ex(aes, s_SSL.Aes128, nullptr, key, iv);
+        rc = EVP_DecryptInit_ex(aes, ssl::Aes128(), nullptr, key, iv);
 #ifdef PODOFO_HAVE_LIBIDN
     else if (keyLen == (int)PdfKeyLength::L256 / 8)
-        rc = EVP_DecryptInit_ex(aes, s_SSL.Aes256, nullptr, key, iv);
+        rc = EVP_DecryptInit_ex(aes, ssl::Aes256(), nullptr, key, iv);
 #endif
     else
         PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InternalLogic, "Invalid AES key length");
@@ -1261,10 +1195,10 @@ void PdfEncryptAESBase::BaseEncrypt(const unsigned char* key, unsigned keyLen, c
 
     int rc;
     if (keyLen == (int)PdfKeyLength::L128 / 8)
-        rc = EVP_EncryptInit_ex(aes, s_SSL.Aes128, nullptr, key, iv);
+        rc = EVP_EncryptInit_ex(aes, ssl::Aes128(), nullptr, key, iv);
 #ifdef PODOFO_HAVE_LIBIDN
     else if (keyLen == (int)PdfKeyLength::L256 / 8)
-        rc = EVP_EncryptInit_ex(aes, s_SSL.Aes256, nullptr, key, iv);
+        rc = EVP_EncryptInit_ex(aes, ssl::Aes256(), nullptr, key, iv);
 #endif
     else
         PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InternalLogic, "Invalid AES key length");
@@ -1477,7 +1411,7 @@ void PdfEncryptSHABase::ComputeHash(const unsigned char* pswd, unsigned pswdLen,
 
     int rc;
     unique_ptr<EVP_MD_CTX, decltype(&EVP_MD_CTX_free)> sha256(EVP_MD_CTX_new(), EVP_MD_CTX_free);
-    if (sha256 == nullptr || (rc = EVP_DigestInit_ex(sha256.get(), s_SSL.SHA256, nullptr)) != 1)
+    if (sha256 == nullptr || (rc = EVP_DigestInit_ex(sha256.get(), ssl::SHA256(), nullptr)) != 1)
         PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InternalLogic, "Error initializing sha256 hashing engine");
 
     if (pswdLen != 0)
@@ -1534,7 +1468,7 @@ void PdfEncryptSHABase::ComputeHash(const unsigned char* pswd, unsigned pswdLen,
             // I'm not 100% sure the conversion is correct, since we don't
             // finalize the context. It may be unecessary because of some
             // preconditions, but these should be clearly stated
-            rc = EVP_EncryptInit_ex(aes.get(), s_SSL.Aes128, nullptr, block, block + 16);
+            rc = EVP_EncryptInit_ex(aes.get(), ssl::Aes128(), nullptr, block, block + 16);
             rc = EVP_EncryptUpdate(aes.get(), data, &dataOutMoved, data, dataLen);
             PODOFO_ASSERT((unsigned)dataOutMoved == dataLen);
 
@@ -1545,19 +1479,19 @@ void PdfEncryptSHABase::ComputeHash(const unsigned char* pswd, unsigned pswdLen,
 
             if (blockLen == 32)
             {
-                rc = EVP_DigestInit_ex(sha256.get(), s_SSL.SHA256, nullptr);
+                rc = EVP_DigestInit_ex(sha256.get(), ssl::SHA256(), nullptr);
                 rc = EVP_DigestUpdate(sha256.get(), data, dataLen);
                 rc = EVP_DigestFinal_ex(sha256.get(), block, nullptr);
             }
             else if (blockLen == 48)
             {
-                rc = EVP_DigestInit_ex(sha384.get(), s_SSL.SHA384, nullptr);
+                rc = EVP_DigestInit_ex(sha384.get(), ssl::SHA384(), nullptr);
                 rc = EVP_DigestUpdate(sha384.get(), data, dataLen);
                 rc = EVP_DigestFinal_ex(sha384.get(), block, nullptr);
             }
             else
             {
-                rc = EVP_DigestInit_ex(sha512.get(), s_SSL.SHA512, nullptr);
+                rc = EVP_DigestInit_ex(sha512.get(), ssl::SHA512(), nullptr);
                 rc = EVP_DigestUpdate(sha512.get(), data, dataLen);
                 rc = EVP_DigestFinal_ex(sha512.get(), block, nullptr);
             }
@@ -1596,7 +1530,7 @@ void PdfEncryptSHABase::ComputeUserKey(const unsigned char* userpswd, unsigned l
 
     int rc;
     unique_ptr<EVP_CIPHER_CTX, decltype(&EVP_CIPHER_CTX_free)> aes(EVP_CIPHER_CTX_new(), EVP_CIPHER_CTX_free);
-    if (aes == nullptr || (rc = EVP_EncryptInit_ex(aes.get(), s_SSL.Aes256, nullptr, hashValue, nullptr)) != 1)
+    if (aes == nullptr || (rc = EVP_EncryptInit_ex(aes.get(), ssl::Aes256(), nullptr, hashValue, nullptr)) != 1)
         PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InternalLogic, "Error initializing AES encryption engine");
 
     EVP_CIPHER_CTX_set_padding(aes.get(), 0); // disable padding
@@ -1640,7 +1574,7 @@ void PdfEncryptSHABase::ComputeOwnerKey(const unsigned char* ownerpswd, unsigned
 
     int rc;
     unique_ptr<EVP_CIPHER_CTX, decltype(&EVP_CIPHER_CTX_free)> aes(EVP_CIPHER_CTX_new(), EVP_CIPHER_CTX_free);
-    if (aes == nullptr || (rc = EVP_EncryptInit_ex(aes.get(), s_SSL.Aes256, nullptr, hashValue, nullptr)) != 1)
+    if (aes == nullptr || (rc = EVP_EncryptInit_ex(aes.get(), ssl::Aes256(), nullptr, hashValue, nullptr)) != 1)
         PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InternalLogic, "Error initializing AES encryption engine");
 
     EVP_CIPHER_CTX_set_padding(aes.get(), 0); // disable padding
@@ -1785,7 +1719,7 @@ void PdfEncryptAESV3::GenerateEncryptionKey(const string_view& documentId)
 
     int rc;
     unique_ptr<EVP_CIPHER_CTX, decltype(&EVP_CIPHER_CTX_free)> aes(EVP_CIPHER_CTX_new(), EVP_CIPHER_CTX_free);
-    if (aes == nullptr || (rc = EVP_EncryptInit_ex(aes.get(), s_SSL.Aes256, nullptr, m_encryptionKey, nullptr)) != 1)
+    if (aes == nullptr || (rc = EVP_EncryptInit_ex(aes.get(), ssl::Aes256(), nullptr, m_encryptionKey, nullptr)) != 1)
         PODOFO_RAISE_ERROR_INFO(PdfErrorCode::InternalLogic, "Error initializing AES encryption engine");
 
     EVP_CIPHER_CTX_set_padding(aes.get(), 0); // disable padding
@@ -1833,7 +1767,7 @@ bool PdfEncryptAESV3::Authenticate(const string_view& password, const string_vie
             // AES-256 in CBC mode with no padding and an initialization vector of zero.
             // The 32-byte result is the file encryption key"
             EVP_CIPHER_CTX* aes = m_aes->getEngine();
-            EVP_DecryptInit_ex(aes, s_SSL.Aes256, nullptr, hashValue, 0); // iv zero
+            EVP_DecryptInit_ex(aes, ssl::Aes256(), nullptr, hashValue, 0); // iv zero
             EVP_CIPHER_CTX_set_padding(aes, 0); // no padding
             int lOutLen;
             EVP_DecryptUpdate(aes, m_encryptionKey, &lOutLen, m_oeValue, 32);
@@ -1850,7 +1784,7 @@ bool PdfEncryptAESV3::Authenticate(const string_view& password, const string_vie
         // AES-256 in CBC mode with no padding and an initialization vector of zero.
         // The 32-byte result is the file encryption key"
         EVP_CIPHER_CTX* aes = m_aes->getEngine();
-        EVP_DecryptInit_ex(aes, s_SSL.Aes256, nullptr, hashValue, 0); // iv zero
+        EVP_DecryptInit_ex(aes, ssl::Aes256(), nullptr, hashValue, 0); // iv zero
         EVP_CIPHER_CTX_set_padding(aes, 0); // no padding
         int lOutLen;
         EVP_DecryptUpdate(aes, m_encryptionKey, &lOutLen, m_ueValue, 32);